分类管理页面完成，歌单歌曲添加未实现

+++ sonic-shelf-server/src/main/java/com/zhongxin/sonicshelf/entity/Categories.java
@@ -4,8 +4,8 @@ import lombok.Data;
 
 @Data
 public class Categories {
-    private Long id;
-    private Long parentId;
+    private Integer id;
+    private Integer parentId;
     private String name;
     private String description;
 }
